requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.49.0
ollama
Pillow>=10.0.0
flask>=3.0.0
waitress>=3.0.0
chainlit>=2.0.0
